move share_precheck models
